# Speedup ~2x hessian_matrix_eigvals and 2D structure_tensor_eigenvalues.