@anntzer 2022-07-16
Thanks for the review.  I realized there was an even faster way to write this, which is also more legible and avoids the needs of `out=...`.  I also addressed the other comments.